+++ src/pages/index.astro
@@ -1,8 +1,8 @@
 ---
 // JS
 import Layout from '../layouts/Main.astro'
-import Hero from '../components/Hero.jsx';
-import Carousel from '../components/Carousel.jsx';
+import Hero from '../components/Hero.tsx';
+import Carousel from '../components/Carousel.tsx';
 import '../styles/carousel.css';
 
 const images = [
